rumor-mill: Disable chunking in GCS table code.
The default chunk size results in high memory usage (kept getting
OOM-killed when deployed on the node).  According to the GCS client
docs [1], it's a good idea to disable chunking (or set chunk size to
upload size) when uploading many small objects, which is what we're
doing.

It appears to vastly reduce memory usage, versus the default 8MiB
buffer size.

[1] https://pkg.go.dev/cloud.google.com/go/storage#Writer