harpoon: In-progress work on scenepack
I suspect one of the things giving the C++ version of ballistae an edge
is heavier optimization on the constant scene built in to the binary.
Loading the scene from a proto file should remove that edge.

The basic idea is to compile from a scene description language (maybe
starlark, eventually) down to a zip file that contains a protobuf scene
description along with all resources.